+++ src/main/java/nbcamp/food_order_platform/store/application/service/StoreService.java
@@ -7,19 +7,18 @@ import nbcamp.food_order_platform.global.error.ErrorCode;
 import nbcamp.food_order_platform.global.error.exception.BusinessException;
 import nbcamp.food_order_platform.regionCode.domain.entity.RegionCode;
 import nbcamp.food_order_platform.regionCode.domain.repository.RegionCodeRepository;
-import nbcamp.food_order_platform.store.application.dto.request.CreateStoreCommand;
-import nbcamp.food_order_platform.store.application.dto.request.GetStoresAdminPageQuery;
-import nbcamp.food_order_platform.store.application.dto.request.GetStoresPageQuery;
-import nbcamp.food_order_platform.store.application.dto.request.UpdateStoreCommand;
-import nbcamp.food_order_platform.store.application.dto.response.GetStoresAdminPageResult;
-import nbcamp.food_order_platform.store.application.dto.response.GetStoresPageResult;
+import nbcamp.food_order_platform.store.application.dto.command.CreateStoreCommand;
+import nbcamp.food_order_platform.store.application.dto.query.GetStoresAdminPageQuery;
+import nbcamp.food_order_platform.store.application.dto.query.GetStoresPageQuery;
+import nbcamp.food_order_platform.store.application.dto.command.UpdateStoreCommand;
+import nbcamp.food_order_platform.store.application.dto.result.GetStoresAdminPageResult;
+import nbcamp.food_order_platform.store.application.dto.result.GetStoresPageResult;
 import nbcamp.food_order_platform.store.application.dto.common.StoreResult;
 import nbcamp.food_order_platform.store.domain.entity.Store;
 import nbcamp.food_order_platform.store.domain.entity.StoreCategory;
 import nbcamp.food_order_platform.store.domain.entity.StoreRegion;
 import nbcamp.food_order_platform.store.domain.repository.StoreRepository;
 import nbcamp.food_order_platform.user.domain.entity.Role;
-import nbcamp.food_order_platform.user.domain.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
