+++ src/main/java/nbcamp/food_order_platform/store/presentation/controller/StoreController.java
@@ -3,13 +3,13 @@ package nbcamp.food_order_platform.store.presentation.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import nbcamp.food_order_platform.global.security.AuthUser;
-import nbcamp.food_order_platform.store.application.dto.request.CreateStoreCommand;
+import nbcamp.food_order_platform.store.application.dto.command.CreateStoreCommand;
 import nbcamp.food_order_platform.store.application.dto.common.StoreResult;
-import nbcamp.food_order_platform.store.application.dto.request.GetStoresAdminPageQuery;
-import nbcamp.food_order_platform.store.application.dto.request.GetStoresPageQuery;
-import nbcamp.food_order_platform.store.application.dto.request.UpdateStoreCommand;
-import nbcamp.food_order_platform.store.application.dto.response.GetStoresAdminPageResult;
-import nbcamp.food_order_platform.store.application.dto.response.GetStoresPageResult;
+import nbcamp.food_order_platform.store.application.dto.query.GetStoresAdminPageQuery;
+import nbcamp.food_order_platform.store.application.dto.query.GetStoresPageQuery;
+import nbcamp.food_order_platform.store.application.dto.command.UpdateStoreCommand;
+import nbcamp.food_order_platform.store.application.dto.result.GetStoresAdminPageResult;
+import nbcamp.food_order_platform.store.application.dto.result.GetStoresPageResult;
 import nbcamp.food_order_platform.store.application.service.StoreService;
 import nbcamp.food_order_platform.store.presentation.dto.common.StoreResDto;
 import nbcamp.food_order_platform.store.presentation.dto.request.PostStoreReqDto;
